api/requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
passlib[bcrypt]